CHORE: move queries to domain layer
