+++ src/hexagonal/adapters/drivens/repository/sqlalchemy/models.py
@@ -45,7 +45,7 @@ def create_aggregates_table(table_name: str, schema: str | None = None) -> Table
         full_table_name,
         metadata,
         Column("originator_id", String(36), nullable=False),
-        Column("aggregate_name", Text, nullable=False),
+        Column("aggregate_name", String(255), nullable=False),
         Column("originator_version", Integer, nullable=False),
         Column("topic", Text, nullable=False),
         Column("state", LargeBinary, nullable=False),
@@ -77,7 +77,7 @@ def create_events_table(table_name: str, schema: str | None = None) -> Table:
         full_table_name,
         metadata,
         Column("originator_id", String(36), nullable=False),
-        Column("aggregate_name", Text, nullable=False),
+        Column("aggregate_name", String(255), nullable=False),
         Column("originator_version", Integer, nullable=False),
         Column("topic", Text, nullable=False),
         Column("state", LargeBinary, nullable=False),
